pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmbase"
version = "0.1.0"
description = "LLM-powered personal knowledge base, inspired by Karpathy"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "flask>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "markdownify>=0.13.0",
    "matplotlib>=3.8.0",
    "python-frontmatter>=1.1.0",
    "watchdog>=4.0.0",
]

[project.scripts]
llmbase = "tools.cli:main"

[tool.setuptools.packages.find]
include = ["tools*"]
